Introduced overlay-class and container-class attributes.
This commit reworks how css classes are provided to the modal-dialog component and then handed down to the overlay div and the position-container component. Previously, we attempted to use one property for both template-provided css classes as well as subclass-provided css classes. That is problematic for the same reason that Ember separates classNames and classNameBindings. After this commit, if you want to specify additional class name(s) when you are using `modal-dialog` in a hbs template, use `overlay-class` and/or `container-class`, and if you want to specify additional class names(s) when you are subclasses the `modal-dialog` component, use `overlayClassNames` and/or `containerClassNames`. The dummy app has been updated with examples.