Respect MODAL_DIALOG_USE_TETHER application property
This provides a mechanism to override the auto detection of
ember-tether. The primary use case for this is to allow
ember-modal-dialog's acceptance tests to test with and without
tether.

We are adding a dev dependency on ember-legacy-views in order to
continue to support setting a view as an alignmentTarget in
Ember 2.0 beta or later. If you are not using this feature, things will
work fine without this dependency being present in your application.