fixes my previous change to the css installation instructions
There are two places in the Readme where the user is told how to install CSS.
The first place is more of a quick start guide and only shows how to import
using regular CSS. The second place shows how to import both CSS and SASS.

As a SASS user installing the addon it is easy to not notice the second section and get tripped
and use the CSS imports instead of the SASS imports, as I did, and never notice the second section.

This commit reverts an earlier change I made and copies the SASS instructions to
the quickstart section.